Refactor import statements to use the new package structure for mappers
