+++ src/main/java/es/marcha/backend/services/ecommerce/ProductService.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Service;
 import es.marcha.backend.dto.response.ecommerce.product.ProductResponseDTO;
 import es.marcha.backend.dto.response.ecommerce.product.ProductReviewResponseDTO;
 import es.marcha.backend.exception.ProductException;
-import es.marcha.backend.mapper.ProductMapper;
+import es.marcha.backend.mapper.product.ProductMapper;
 import es.marcha.backend.model.ecommerce.product.Product;
 import es.marcha.backend.repository.ecommerce.ProductRepository;
 import es.marcha.backend.utils.ProductUtils;
