+++ src/main/java/es/marcha/backend/services/ecommerce/SubcategoryService.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Service;
 
 import es.marcha.backend.dto.response.ecommerce.SubcategoryResponseDTO;
 import es.marcha.backend.exception.ProductException;
-import es.marcha.backend.mapper.SubcategoryMapper;
+import es.marcha.backend.mapper.product.SubcategoryMapper;
 import es.marcha.backend.model.ecommerce.Subcategory;
 import es.marcha.backend.repository.ecommerce.SubcategoryRepository;
 import es.marcha.backend.utils.ProductUtils;
