+++ src/main/java/es/marcha/backend/services/order/OrderAddressService.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Service;
 
 import es.marcha.backend.dto.response.order.OrderAddrResponseDTO;
 import es.marcha.backend.exception.OrderException;
-import es.marcha.backend.mapper.OrderAddrMapper;
+import es.marcha.backend.mapper.order.OrderAddrMapper;
 import es.marcha.backend.model.order.OrderAddresses;
 import es.marcha.backend.repository.order.OrderAddrRepository;
 import jakarta.transaction.Transactional;
