+++ src/main/java/es/marcha/backend/services/order/OrderItemsService.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Service;
 
 import es.marcha.backend.dto.response.order.OrderItemsResponseDTO;
 import es.marcha.backend.exception.OrderException;
-import es.marcha.backend.mapper.OrderItemMapper;
+import es.marcha.backend.mapper.order.OrderItemMapper;
 import es.marcha.backend.model.order.OrderItems;
 import es.marcha.backend.repository.order.OrderItemRepository;
 import jakarta.transaction.Transactional;
