+++ src/main/java/es/marcha/backend/services/order/OrderService.java
@@ -9,8 +9,8 @@ import org.springframework.stereotype.Service;
 import es.marcha.backend.dto.response.order.OrderAddrResponseDTO;
 import es.marcha.backend.dto.response.order.OrderResponseDTO;
 import es.marcha.backend.exception.OrderException;
-import es.marcha.backend.mapper.OrderAddrMapper;
-import es.marcha.backend.mapper.OrderMapper;
+import es.marcha.backend.mapper.order.OrderAddrMapper;
+import es.marcha.backend.mapper.order.OrderMapper;
 import es.marcha.backend.model.enums.OrderStatus;
 import es.marcha.backend.model.order.Order;
 import es.marcha.backend.model.user.Address;
